// src/models/seach-client-settings.ts
export class SearchClientSettingBy {
    id: number;
    displayName: string;
    name: string;
    data: string;
    status: number;
    createdBy: number;
    createdDate: string;
    updatedBy: number;
    updatedDate: string;
    deletedBy: number;
    deletedDate: string;
    startDate: string;
    endDate: string;
    names: string[];
    nameLike: string;
}